feat: Extract controls into separate files
